requirements.txt
llama-index-llms-google-genai 
google-genai